+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 1.1.0
+- Support for React 16
+- Use prop-types module instead of React.propTypes
+- Use "import" instead of "require"
+
 ## 1.0.3
 - Display Google maps address by default if there is no input user address